+++ packages/storebuilder/src/wizards/store-setup/data/constants.ts
@@ -23,17 +23,13 @@ export const StoreSetupStringData = {
 		productTypesLabelText: __('What types of products are you selling?', 'nexcess-mapps'),
 		productTypesHelperText: __('Select all that apply.', 'nexcess-mapps'),
 		productCountLabelText: __('How many products do you have?', 'nexcess-mapps')
-
 	},
 	complete: {
 		title: __('Nice work! Let\'s keep going.', 'nexcess-mapps'),
-		copy: __(
-			'Nam erat felis, consectetur et velit non, fermentum vulputate sapien. Morbi rhoncus a metus et.',
-			'nexcess-mapps'
-		),
+		copy: __('Next up: Add products and set up your payment method.', 'nexcess-mapps')
 	},
 	submitForm: {
-		errorMessage: __('There was an error saving the data.', 'nexcess-mapps'),
+		errorMessage: __('There was an error saving the data.', 'nexcess-mapps')
 	}
 };
 
@@ -42,20 +38,20 @@ export const productTypeOptions = [
 		value: 'physical-goods',
 		label: __('Physical Goods', 'nexcess-mapps'),
 		icon: `${ IMAGE_DIR }product-type-physical.png`,
-		name: 'product_type',
+		name: 'product_type'
 	},
 	{
 		value: 'digital-goods',
 		label: __('Digital Goods', 'nexcess-mapps'),
 		icon: `${ IMAGE_DIR }product-type-digital.png`,
-		name: 'product_type',
+		name: 'product_type'
 	},
 	{
 		value: 'services',
 		label: __('Services', 'nexcess-mapps'),
 		icon: `${ IMAGE_DIR }product-type-services.png`,
-		name: 'product_type',
-	},
+		name: 'product_type'
+	}
 ];
 
 export const productCountOptions = [
@@ -63,18 +59,18 @@ export const productCountOptions = [
 		value: '1-10',
 		label: '1-10',
 		icon: `${ IMAGE_DIR }ftc-product-count-1-10.svg`,
-		name: 'product_count',
+		name: 'product_count'
 	},
 	{
 		value: '10-30',
 		label: '10-30',
 		icon: `${ IMAGE_DIR }ftc-product-count-10-30.svg`,
-		name: 'product_count',
+		name: 'product_count'
 	},
 	{
 		value: '30-99',
 		label: '30-99+',
 		icon: `${ IMAGE_DIR }ftc-product-count-30-99.svg`,
-		name: 'product_count',
-	},
+		name: 'product_count'
+	}
 ];